move BulkPatch out of HydraModel to API
